requirements.txt
aiogram==3.4.1
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
alembic==1.16.1
pyjwt==2.10.1
python-multipart==0.0.9
jinja2==3.1.3
python-dotenv==1.0.1
python-telegram-bot==20.6
aiofiles==23.2.1
aiohttp==3.9.3
psycopg2-binary==2.9.9
pydantic==2.6.1 